fix: make sure not to return error, when not finding user

+++ app/Http/Controllers/E2E/UserController.php
@@ -18,8 +18,12 @@ class UserController extends Controller
   {
     if($request->has('phone_number'))
     {
-      $user = User::wherePhoneNumber($request->phone_number)->firstOrFail();
-      $user->delete();
+      $user = User::wherePhoneNumber($request->phone_number)->first();
+      
+      if($user)
+      {
+        $user->delete();
+      }
     }
   }
